fix(core): correct GenerationOptions.temperature type from u32 to f32
Temperature is conventionally a float; the placeholder u32 was flagged
during W18 retro by reviewers (senior orientation + Rick during M1.8
prep). Pure type change in the GenerationOptions struct.

Drops the Eq derive from GenerationOptions and from runtime's
GenerateRequest because f32 is not Eq. PartialEq is preserved.

Refs: M1.6, W18 retro cleanup

+++ crates/core/src/lib.rs
@@ -434,10 +434,10 @@ pub struct ModelMetadata {
     pub tokenizer_model_hint: Option<String>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct GenerationOptions {
     pub max_new_tokens: usize,
-    pub temperature: Option<u32>,
+    pub temperature: Option<f32>,
 }
 
 impl Default for GenerationOptions {
